requirements.txt
mcp==1.9.3
openai==1.85.0
python-dotenv==1.1.0
langchain==0.3.25
langchain-community==0.3.25
langchain-chroma==0.2.4
langchain-openai==0.3.23
pypdf==5.6.0
